Moved Deezer Metadata fetch into its own file and added better song handling
