+++ src/getSongLink/getSongLink.ts
@@ -1,13 +1,13 @@
 import { getLinks } from "songlink-api";
-import * as deezerApi from "deezer-api-ts";
 import { SonglinkPlatform } from "songlink-api/lib/types/Platform";
 import { SonglinkResponse } from "songlink-api/lib/types/Response";
 import { LoggerLayers } from "../config";
 import { createLayeredLogger } from "../utils/logger";
+import { getDeezerMetadata } from "./metadata/deezer";
 
 const logger = createLayeredLogger(LoggerLayers.songLink);
 
-interface MetaData {
+export interface MetaData {
 	title?: string;
 	artistName?: string;
 	artwork?: string;
@@ -40,41 +40,7 @@ export const getIdFromEntity = (entity: string): string => {
 
 export const getDetailedMetadata = async (data: SonglinkResponse): Promise<MetaData> => {
 	if (data.linksByPlatform.deezer) {
-		try {
-			logger.info("Fetching metadata from Deezer");
-
-			const entityId = getIdFromEntity(data.linksByPlatform.deezer.entityUniqueId);
-
-			let metaData;
-			if (data.linksByPlatform.deezer.entityUniqueId.includes("SONG")) {
-				const songMetaData = await deezerApi.getTrack(parseInt(entityId));
-				metaData = {
-					title: songMetaData.title,
-					artist: songMetaData.artist,
-					artwork: songMetaData.album?.cover || ""
-				};
-			} else {
-				metaData = await deezerApi.getAlbum(parseInt(entityId));
-			}
-
-
-			if (!metaData.title) {
-				logger.warn("Attempted to fetch metadata from deezer but received empty response");
-				return;
-			}
-
-			logger.info(`Successfully fetched metadata for ${metaData.title}`);
-
-			return {
-				title: metaData.title,
-				artistName: metaData.artist.name,
-				artwork: metaData.cover,
-				genres: metaData.genres?.data.map(genre => genre.name),
-				releaseYear: metaData.release_date
-			};
-		} catch (e) {
-			logger.error("Unable to fetch metadata from Deezer:", e);
-		}
+		return await getDeezerMetadata(data);
 	} else {
 		logger.warn("Requested metaData but it is not on any supported streaming services");
 	}
